feat: implement role-based access control for category, comment, post, and tag management

+++ src/models/commentModel.ts
@@ -1,4 +1,5 @@
 import prisma from '../config/database.js';
+import { PrismaClient, UserRole } from '@prisma/client';
 
 export const getAllComments = (commenter?: string, post?: number) => {
     return prisma.comment.findMany({
@@ -14,22 +15,66 @@ export const getCommentById = (id: number) => {
     });
 }
 
-export const createComment = (data: { post_id: number, commenter_name: string, content: string }) => {
+export const createComment = (data: { post_id: number, commenter_name: string, content: string }, currentUser: { id: number, role: UserRole }) => {
+    if (!currentUser) {
+        throw new Error('Unauthorized');
+    }
     return prisma.comment.create({
-        data
+        data: {
+            ...data,
+            user_id: currentUser.id
+        }
     });
 }
 
 
-export const updateComment = (id: number, data: object) => {
+export const updateComment = async (id: number, data: object, currentUser: { id: number, role: UserRole }) => {
+
+    const commentItem = await prisma.comment.findUnique({
+        where: { id },
+        select: {
+            user_id: true
+        }
+    });
+
+    if (!commentItem) {
+        throw new Error('Comment not found');
+    }
+
+    if (commentItem.user_id !== currentUser.id) {
+        throw new Error('Forbidden');
+    }
+
     return prisma.comment.update({
         where: { id },
         data: data
     });
 }
 
-export const deleteComment = (id: number) => {
-    return prisma.comment.delete({
-        where: { id }
+export const deleteComment = async (id: number, currentUser: { id: number, role: UserRole }) => {
+    const commentItem = await prisma.comment.findUnique({
+        where: { id },
+        include: {
+            post: {
+                select: {
+                    user_id: true,
+                },
+            },
+        }
+    });
+
+    if (!commentItem) {
+        throw new Error('Comment not found');
+    }
+
+    const isCommenter = commentItem.user_id === currentUser.id;
+    const isPostOwner = commentItem.post.user_id === currentUser.id;
+    const isAdminOrModerator = currentUser.role === 'ADMIN' || currentUser.role === 'MODERATOR';
+    if (!isCommenter && !isPostOwner && !isAdminOrModerator) {
+        return prisma.comment.delete({
+            where: { id },
         });
+    } else {
+        throw new Error('Forbidden');
+    }
 }
