+++ src/models/tagModel.ts
@@ -1,6 +1,5 @@
-import { Prisma } from '@prisma/client';
+import { Prisma, UserRole } from '@prisma/client';
 import prisma from '../config/database.js';
-import { SHOW_DELETED_OPTIONS, POST_STATUS } from '../constants.js';
 
 export const getAllTags = () => {
     return prisma.tag.findMany();
@@ -12,20 +11,29 @@ export const getTagById = (id: number) => {
     });
 };
 
-export const createTag = (body: Prisma.TagCreateInput) => {
+export const createTag = (body: Prisma.TagCreateInput, currentUser: {id : number, role: UserRole}) => {
+    if(currentUser.role !== 'ADMIN'){
+        throw new Error('Forbidden');
+    }
     return prisma.tag.create({
         data: body
     });
 };
 
-export const updateTag = (id: number, data: Prisma.TagUpdateInput) => {
+export const updateTag = (id: number, data: Prisma.TagUpdateInput, currentUser: {id : number, role: UserRole}) => {
+    if(currentUser.role !== 'ADMIN'){
+        throw new Error('Forbidden');
+    }
     return prisma.tag.update({
         where: { id },
         data,
     });
 };
 
-export const deleteTag = (id: number) => {
+export const deleteTag = (id: number, currentUser: {id : number, role: UserRole}) => {
+    if(currentUser.role !== 'ADMIN'){
+        throw new Error('Forbidden');
+    }
     return prisma.tag.delete({
         where: { id }
     });
